% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumns.R
\name{addList}
\alias{addList}
\title{Add list of new columns to a table}
\usage{
x <- dplyr::data_frame(a=1:5, b=a+7)
l <- list(
  c("col_1", "a + b"),
  c("col_2_name", ~pmax(a*a, a*b - a*a))
)
y <- addList(x, l)
}
\arguments{
\item{table}{table to which new column should be added}

\item{funcList}{list containing pairs of new column name and new column lazy-evaluable value}
}
\value{
table with added new columns
}
\description{
Add list of new columns to a table
}

