% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumns.R
\name{addSingle}
\alias{addSingle}
\title{Add single new column to table}
\usage{
addSingle(table, nameFunct)
}
\arguments{
\item{table}{table to which new column should be added}

\item{nameFunct}{pair (newColumnName, newColumnValue)}
}
\value{
table with added new column
}
\description{
Add single new column to table
}

