% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyznaczniki.R
\docType{data}
\name{all_rozsz}
\alias{all_rozsz}
\title{String zawierający wszystkie matury podstawowe dostępne w serwisie}
\format{An object of class \code{character} of length 1.}
\usage{
all_rozsz
}
\description{
String zawierający wszystkie matury podstawowe dostępne w serwisie
}
\keyword{datasets}

