% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj.kolumny.R
\name{dodaj.kolumny}
\alias{dodaj.kolumny}
\title{Dodaj nowe kolumny do zbioru danych}
\usage{
dodaj.kolumny(ramka, lista.funkcji)
}
\arguments{
\item{ramka}{data.frame do którego należy dodać nowe kolumny}

\item{lista.funkcji}{lista zawierająca pary postaci
("nazwa_nowej_kolumny", "formuła_wyliczana_dla_każdego_wiersza")}
}
\value{
ramka wraz z dodanymi nowymi kolumnami
}
\description{
Dodaj nowe kolumny do zbioru danych
}
\examples{
{
x <- dplyr::data_frame(a=1:5, b=a+7)
l <- list(
  c("col_1", "a + b"),
  c("col_2_name", ~pmax(a*a, a*b - a*a))
)
y <- dodaj.kolumny(ramka = x, lista.funkcji = l)
}
}

