% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generuj.wykres.porownaj.R
\name{generuj.wykres.porownaj}
\alias{generuj.wykres.porownaj}
\title{Generuj wykresy porównawcze dla konkretnego współczynnika dla dwóch szkół}
\usage{
generuj.wykres.porownaj(dane, nazwa.kolumny, szkola1_id, szkola2_id, opis)
}
\arguments{
\item{dane}{tabela z której generować wykres}

\item{nazwa.kolumny}{nazwa kolumny}

\item{szkola1_id}{id pierwszej szkoły dla której ma być wygenerowany zestaw danych w ramach tego wykresu}

\item{szkola2_id}{id drugiej szkoły dla której ma być wygenerowany zestaw danych w ramach tego wykresu}

\item{opis}{nazwa dla danych zawartych w wyszczególnionej kolumnie - tytuł wykresu}
}
\description{
Dla współczynnika o wartościach liczbowych generuje wykres liniowy gęstości
Dla współczynnika o wartościach innych (tekst) generuje słupkowy wykres gęstościowy
}
\examples{
{
generuj.wykres(matura.2015, "m_mat_r", 24063, 24064, "Matura rozszerzona z matematyki")
}
}

