% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generuj.wykres.R
\name{generuj.wykres}
\alias{generuj.wykres}
\title{Generuj wykresy porównawcze dla konkretnego współczynnika}
\usage{
generuj.wykres(dane, nazwa.kolumny, szkola_id, opis)
}
\arguments{
\item{dane}{tabela z której generować wykres}

\item{nazwa.kolumny}{nazwa kolumny}
}
\description{
Generuj wykresy porównawcze dla konkretnego współczynnika
}

