% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matura.2015.R
\docType{data}
\name{matura.2015}
\alias{matura.2015}
\title{Matura results from 2015}
\format{a data frame with 360432 rows and 16 columns}
\usage{
data(matura.2015)
}
\description{
Pulled from http://zpd.ibe.edu.pl
}
\keyword{datasets}

