% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{maturiser.server}
\alias{maturiser.server}
\title{Serwer aplikacji shiny}
\usage{
maturiser.server(matury, wyznaczniki)
}
\arguments{
\item{matury}{tabela zawierająca wyniki matur i wyliczonych wyznaczników}

\item{wyznaczniki}{lista par postaci ("nazwa_kolumny", "opis_wyznacznika")
Wymienione wyznaczniki zostaną przedstawione w aplikacji}
}
\description{
Serwer aplikacji shiny
}

