% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{maturiser.ui}
\alias{maturiser.ui}
\title{UI aplikacji shiny}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
maturiser.ui
}
\description{
UI aplikacji shiny
}
\keyword{datasets}

