% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uruchom.R
\name{uruchom}
\alias{uruchom}
\title{Uruchom aplikację shiny dla konkretynych danych i wyznaczników}
\usage{
uruchom(dane, lista.wyznacznikow)
}
\arguments{
\item{dane}{tabela zawierajaca wyniki matur i wartości
wyznaczników dla poszczególnych uczniów. dla każdego ucznia jest
też id jego szkoły, skrót nazwy gminy i nazwa szkoły}

\item{lista.wyznaczników}{lista wyznaczników które mają być
przedstawione na wykresach. typu c("nazwa.kolumny", "opis.wyznacznika")
gdzie nazwa.kolumny to kolumna w tabeli dane}
}
\description{
Uruchom aplikację shiny dla konkretynych danych i wyznaczników
}

