% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyznaczniki.R
\docType{data}
\name{uw.informatyka}
\alias{uw.informatyka}
\title{Krotka zawierająca opis funkcji wyliczjącej punkty przy rekrutacji na informatykę UW}
\format{An object of class \code{list} of length 3.}
\usage{
uw.informatyka
}
\description{
Krotka zawierająca opis funkcji wyliczjącej punkty przy rekrutacji na informatykę UW
}
\keyword{datasets}

