% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyznaczniki.R
\docType{data}
\name{uw.matematyka}
\alias{uw.matematyka}
\title{Krotka zawierająca opis funkcji wyliczającej punkty przy rekrutacji na matematykę na UW}
\format{An object of class \code{list} of length 3.}
\usage{
uw.matematyka
}
\description{
Krotka zawierająca opis funkcji wyliczającej punkty przy rekrutacji na matematykę na UW
}
\keyword{datasets}

