% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyznaczniki.R
\docType{data}
\name{uw.prawo}
\alias{uw.prawo}
\title{Krotka zawierająca opis funkcji wyliczjącej punkty przy rekrutacji na PRAWO UW}
\format{An object of class \code{list} of length 3.}
\usage{
uw.prawo
}
\description{
Krotka zawierająca opis funkcji wyliczjącej punkty przy rekrutacji na PRAWO UW
}
\keyword{datasets}

