% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyznaczniki.R
\docType{data}
\name{zdawalnosc}
\alias{zdawalnosc}
\title{Krotka zawierająca opis funkcji wyliczającej czy osoba zdała maturę}
\format{An object of class \code{list} of length 3.}
\usage{
zdawalnosc
}
\description{
Krotka zawierająca opis funkcji wyliczającej czy osoba zdała maturę
}
\keyword{datasets}

