% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generuj.wykres.R
\name{generuj.wykres}
\alias{generuj.wykres}
\title{Generuj wykresy porównawcze dla konkretnego współczynnika}
\usage{
generuj.wykres(dane, nazwa.kolumny, szkola_id, opis)
}
\arguments{
\item{dane}{tabela z której generować wykres}

\item{nazwa.kolumny}{nazwa kolumny}

\item{szkola_id}{id szkoły dla której ma być wygenerowany zestaw danych w ramach tego wykresu}

\item{opis}{nazwa dla danych zawartych w wyszczególnionej kolumnie - tytuł wykresu}
}
\description{
Generuj wykresy porównawcze dla konkretnego współczynnika
}
\examples{
{
generuj.wykres(matura.2015, "m_mat_r", 24063, "Matura rozszerzona z matematyki")
}
}

