% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matura.2015.R
\docType{data}
\name{matura.2015}
\alias{matura.2015}
\title{Wyniki egzaminu maturalnego z 2015 roku}
\format{data.frame 360992 x 24
\itemize{
\item id_cke - wewnętrzny identyfikator ucznia 
\item id_szkoly - wewnętrzny identyfikator szkoły
\item nazwa_szkoly - nazwa szkoły
\item gmina_szkoly - nazwa gminy w której znajduje sie szkoła
\item m_mat_p - procentowy wynik z matury podstawowej z matematyki
\item m_pol_p - procentowy wynik z matury podstawowej z j. polskiego
\item m_ang_p - procentowy wynik z matury podstawowej z j. angielskiego
\item m_fiz_p - procentowy wynik z matury podstawowej z fizyki
\item m_bio_p - procentowy wynik z matury podstawowej z biologii
\item m_che_p - procentowy wynik z matury podstawowej z chemii
\item m_inf_p - procentowy wynik z matury podstawowej z informatyki
\item m_wos_p - procentowy wynik z matury podstawowej z wiedzy o społeczeństwie
\item m_his_p - procentowy wynik z matury podstawowej z historii
\item m_geo_p - procentowy wynik z matury podstawowej z geografii
\item m_mat_r - procentowy wynik z matury rozszerzonej z matematyki
\item m_pol_r - procentowy wynik z matury rozszerzonej z j. polskiego
\item m_ang_r - procentowy wynik z matury rozszerzonej z j. angielskiego
\item m_fiz_r - procentowy wynik z matury rozszerzonej z fizyki
\item m_bio_r - procentowy wynik z matury rozszerzonej z biologii
\item m_che_r - procentowy wynik z matury rozszerzonej z chemii
\item m_inf_r - procentowy wynik z matury rozszerzonej z informatyki
\item m_wos_r - procentowy wynik z matury rozszerzonej z wiedzy o społeczeństwie
\item m_his_r - procentowy wynik z matury rozszerzonej z historii
\item m_geo_r - procentowy wynik z matury rozszerzonej z geografii
}}
\usage{
data(matura.2015)
}
\description{
Wyniki egzaminu maturalnego z 2015 roku
}
\keyword{datasets}

