% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyznaczniki.R
\docType{data}
\name{uw.dziennikarstwo}
\alias{uw.dziennikarstwo}
\title{Krotka zawierająca opis funkcji wyliczjącej punkty przy rekrutacji na dziennikarstwo UW}
\format{An object of class \code{list} of length 3.}
\usage{
uw.dziennikarstwo
}
\description{
Krotka zawierająca opis funkcji wyliczjącej punkty przy rekrutacji na dziennikarstwo UW
}
\keyword{datasets}

